refactor: support ordinals with ledger
