+++ src/app/components/loaders/stacks-account-loader.tsx
@@ -4,7 +4,6 @@ import { StacksAccount } from '@app/store/accounts/blockchain/stacks/stacks-acco
 interface CurrentStacksAccountLoaderProps {
   children(data: StacksAccount): React.ReactNode;
 }
-
 export function CurrentStacksAccountLoader({ children }: CurrentStacksAccountLoaderProps) {
   const currentAccount = useCurrentStacksAccount();
   if (!currentAccount) return null;
